Remove dependency on DiscordSRV plugin.
This commit shades in the JDA library that DiscordSRV was using and makes some minor changes to the functions. This avoids a more complicated setup if you wanted to implement all your chat features in CommandHelper. However, if you still want to use DiscordSRV, I recommend you fork 0.0.1 and call it CHDiscordSRV. If you wish to use this new extension, then you'll have to use the discord_connect(token, server_id) function to connect to your server before the functions and events will work again. You'll also have to strip colors yourself as previous DiscordSRV was doing that.