Add support for multiple guild servers
No changes to your code should be required when connected to a single guild. When a guild server is unspecified in a function, it will fall back to event bind context if available, or the default guild server specified during connection.

Also, when not specifying a channel in discord_broadcast(), the function will now use the event bind channel context if available (discord_message_received, discord_voice_joined, and discord_voice_left), or the default channel for the guild. (top most publicly viewable channel)